chore(delete_remake_migrations): use migration framework to detect and remove replaces
The current way of parsing the file can break easily, instead
this commit uses the Django migration framework to detect
and also remove the replaces.
